requirements.txt
# No external dependencies required for the minimal application 
google-cloud-logging==3.5.0
google-cloud-storage==2.9.0
google-cloud-kms==2.18.0
requests==2.31.0
flask==2.3.3
google-cloud-aiplatform==1.36.0
google-generativeai==0.3.1 
# web3>=6.0.0 includes ExtraDataToPOAMiddleware
web3>=6.0.0
python-dotenv==1.0.0 
# The following dependencies are no longer needed as they're included in web3>=6.0.0
# eth-tester>=0.9.0
# py-evm>=0.7.0
# eth-bloom>=1.0.4 